Chrome: Move keystore methods to linux_keystore.py
